Shift Motor IDs to account for no Motor ID 0
With the firmware 0.9.8 our thrusters came with, setting a thruster to
motor ID 0 seems to really mess up communications with it.  This is
probably due to a bug in the thruster firmware.

Also fix CRC problem due to it being statically initialized rather than
it being reset for each new packet.